Clean up function parameters, improve interface.
This introduces the Stacking class, which tracks how the input images / video should be stacked in the output, and consolidates the parameters rows, cols, and mode for all functions.

Additionally, image dimensions should be specified together, and not separately - the parameters width and height have been merged into size.

Finally, the program would print that the output extension for videos was being set to the input extensions, which is now fixed.