Clean up __main__.py, fill in params in docs
Each function which generates images or videos as output does so in a standard pattern, taking in the input files, and then the specific file output, and then named arguments which provide defaults. Since this pattern exists, grouping the args together saves space, and prevents possible issues with incorrect ordering.

Additionally, this commit finishes filling in parameters in existing docs - however, no new docs are added.